+++ modules/actor/cpr-black-ice.js
@@ -1,4 +1,4 @@
-/* globals Actor, duplicate, setProperty */
+/* globals Actor, duplicate, setProperty, game */
 import * as CPRRolls from "../rolls/cpr-rolls.js";
 import CPR from "../system/config.js";
 import LOGGER from "../utils/cpr-logger.js";
@@ -25,7 +25,58 @@ export default class CPRBlackIceActor extends Actor {
     LOGGER.trace("createStatRoll | CPRBlackIceActor | called.");
     const niceStatName = SystemUtils.Localize(CPR.blackIceStatList[statName]);
     const statValue = parseInt(this.data.data.stats[statName], 10);
-    return new CPRRolls.CPRStatRoll(niceStatName, statValue);
+    const cprRoll = new CPRRolls.CPRStatRoll(niceStatName, statValue);
+    if (this.isToken && typeof this.token.data.flags["cyberpunk-red-core"] !== "undefined") {
+      const cprFlags = this.token.data.flags["cyberpunk-red-core"];
+      if (typeof cprFlags.program !== "undefined") {
+        cprRoll.rollCardExtraArgs.program = duplicate(cprFlags.program);
+      }
+    }
+
+    if (cprRoll.rollCardExtraArgs.length === 0) {
+      cprRoll.rollCardExtraArgs.program = {
+        data: {
+          class: "blackice",
+          blackIceType: this.data.data.class,
+        },
+      };
+    }
+    return cprRoll;
+  }
+
+  createDamageRoll(programId, netrunnerTokenId, sceneId) {
+    let program;
+    if (netrunnerTokenId) {
+      const sceneList = (sceneId) ? game.scenes.filter((s) => s.id === sceneId) : game.scenes;
+      let netrunnerToken;
+      sceneList.forEach((scene) => {
+        const tokenList = scene.tokens.filter((t) => t.id === netrunnerTokenId);
+        if (tokenList.length === 1) {
+          [netrunnerToken] = tokenList;
+        }
+      });
+      if (netrunnerToken) {
+        program = netrunnerToken.actor._getOwnedItem(programId);
+      }
+    } else {
+      const programList = game.items.filter((i) => i.data._id === programId);
+      if (programList.length === 1) {
+        [program] = programList;
+      }
+    }
+
+    let damageFormula = "1d6";
+    let programName = this.name;
+    let programData = {};
+    if (program) {
+      damageFormula = (this.data.data.class === "antiprogram") ? program.data.data.damage.blackIce : program.data.data.damage.standard;
+      programName = program.name;
+      programData = program.data;
+    }
+
+    const cprRoll = new CPRRolls.CPRDamageRoll(programName, damageFormula, "program");
+    cprRoll.rollCardExtraArgs.program = programData;
+    return cprRoll;
   }
 
   /**