+++ modules/dialog/cpr-configure-bi-actor-from-program.js
@@ -1,0 +1,44 @@
+/* eslint-disable no-shadow */
+/* global renderTemplate Dialog FormDataExtended */
+import LOGGER from "../utils/cpr-logger.js";
+import SystemUtils from "../utils/cpr-systemUtils.js";
+
+export default class ConfigureBIActorFromProgramPrompt {
+  static async RenderPrompt(data) {
+    const template = "systems/cyberpunk-red-core/templates/dialog/cpr-configure-bi-actor-from-program-prompt.hbs";
+    return new Promise((resolve, reject) => {
+      renderTemplate(template, data).then((html) => {
+        const _onCancel = () => {
+          LOGGER.trace("_onCancel | Dialog ConfigureBIActorFromProgramPrompt | called.");
+          reject();
+        };
+        const _onConfirm = (html) => {
+          LOGGER.trace("_onConfirm | Dialog ConfigureBIActorFromProgramPrompt | called.");
+          const formData = new FormDataExtended(html.find("form")[0]).toObject();
+          resolve(formData);
+        };
+        new Dialog({
+          title: SystemUtils.Localize("CPR.configurefromprogramtitle"),
+          content: html,
+          buttons: {
+            cancel: {
+              icon: "<i class=\"fas fa-times\"></i>",
+              label: SystemUtils.Localize("CPR.cancel"),
+              callback: (html) => _onCancel(html), // TODO fix no-shadow
+            },
+            confirm: {
+              icon: "<i class=\"fas fa-check\"></i>",
+              label: SystemUtils.Localize("CPR.confirm"),
+              callback: (html) => _onConfirm(html), // TODO fix no-shadow
+            },
+          },
+          default: "confirm",
+          render: LOGGER.trace("confirm | Dialog ConfigureBIActorFromProgramPrompt | called."),
+          close: () => {
+            reject();
+          },
+        }).render(true);
+      });
+    });
+  }
+}