+++ modules/rolls/cpr-rolls.js
@@ -137,6 +137,7 @@ export class CPRStatRoll extends CPRRoll {
   constructor(name, value) {
     super(name, "1d10");
     LOGGER.trace(`CPRStatRoll | Constructor`);
+    this.statName = name;
     this.statValue = value;
     this.rollPrompt = "systems/cyberpunk-red-core/templates/dialog/rolls/cpr-verify-roll-stat-prompt.hbs";
     this.rollCard = "systems/cyberpunk-red-core/templates/chat/cpr-stat-rollcard.hbs";
@@ -146,9 +147,9 @@ export class CPRStatRoll extends CPRRoll {
     return this.initialRoll + this.totalMods() + this.statValue;
   }
 
-  setNetCombat() {
-    this.rollPrompt = "systems/cyberpunk-red-core/templates/dialog/rolls/cpr-verify-program-defense-prompt.hbs";
-    this.rollCard = "systems/cyberpunk-red-core/templates/chat/cpr-program-defense-rollcard.hbs";
+  setNetCombat(rollTitle) {
+    this.rollTitle = rollTitle;
+    this.rollCard = "systems/cyberpunk-red-core/templates/chat/cpr-program-stat-rollcard.hbs";
   }
 }
 
@@ -192,7 +193,8 @@ export class CPRAttackRoll extends CPRSkillRoll {
     this.weaponType = weaponType;
   }
 
-  setNetCombat() {
+  setNetCombat(rollTitle) {
+    this.rollTitle = rollTitle;
     this.rollPrompt = "systems/cyberpunk-red-core/templates/dialog/rolls/cpr-verify-program-attack-prompt.hbs";
     this.rollCard = "systems/cyberpunk-red-core/templates/chat/cpr-program-attack-rollcard.hbs";
   }
@@ -333,7 +335,8 @@ export class CPRDamageRoll extends CPRRoll {
     }
   }
 
-  setNetCombat() {
+  setNetCombat(rollTitle) {
+    this.rollTitle = rollTitle;
     this.rollPrompt = "systems/cyberpunk-red-core/templates/dialog/rolls/cpr-verify-program-damage-prompt.hbs";
     this.rollCard = "systems/cyberpunk-red-core/templates/chat/cpr-program-damage-rollcard.hbs";
   }